+++ src/serlo-editor/editor-ui/exercises/interactive-answer.tsx
