+++ src/serlo-editor/editor-ui/index.ts
@@ -1,7 +1,7 @@
 export * from './editor-input'
 export * from './editor-textarea'
 export * from './hover-overlay'
-export * from './exercises/interactive-answer-component'
+export * from './exercises/interactive-answer'
 export * from './exercises/add-button'
 export * from './preview-overlay'
 export * from './preview-overlay-simple'
